feat: add fov(wip)

+++ src/components/field_of_vision.rs
@@ -4,7 +4,7 @@ use bevy::prelude::*;
 #[derive(Component)]
 pub struct FieldOfVision {
     pub visible_tiles: Vec<Position>,
-    pub range: u64,
+    pub range: i32,
     /// flag indicate the field of view changed.
     pub invalid: bool,
 }
